# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePredictionResultFeedbackRequest": {
        "properties": {
          "order": {
            "format": "int32",
            "type": "integer"
          },
          "resultId": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "enum": [
              "OUTPUT",
              "EXPLANATION"
            ],
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "required": [
          "order",
          "resultId",
          "type",
          "value"
        ],
        "type": "object"
      },
...
      "JsonNode": { },
...
      "PageDtoPluginDto": {
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PluginDto"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalItems": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "hasNext",
          "items",
          "page",
          "size",
          "totalItems"
        ],
        "type": "object"
      },
...
      "PluginDto": {
        "properties": {
          "contentType": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "pluginType": {
            "type": "string"
          },
          "sizeBytes": {
            "format": "int64",
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedByAvatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByName": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "contentType",
          "createdAt",
          "fileName",
          "id",
          "kind",
          "pluginType",
          "sizeBytes",
          "source",
          "updatedAt",
          "updatedByAvatarUrl",
          "updatedByEmail",
          "updatedByName"
        ],
        "type": "object"
      },
...
      "PredictionResultFeedbackDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "order": {
            "format": "int32",
            "type": "integer"
          },
          "resultId": {
            "format": "int64",
            "type": "integer"
          },
          "type": {
            "enum": [
              "OUTPUT",
              "EXPLANATION"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "userId": {
            "format": "int64",
            "type": "integer"
          },
          "userName": {
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "required": [
          "createdAt",
          "id",
          "order",
          "resultId",
          "type",
          "updatedAt",
          "userEmail",
          "userId",
          "userName",
          "value"
        ],
        "type": "object"
      },
...
      "ServiceActionRequest": {
        "properties": {
          "action": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "type": "object"
      },
...
      "TerminalSessionRequest": {
        "properties": {
          "cols": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "rows": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "serviceName": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "cols",
          "rows",
          "serviceName"
        ],
        "type": "object"
      },
      "TerminalSessionResponse": {
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "wsPath": {
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "wsPath"
        ],
        "type": "object"
      },
...
      "UpdatePredictionResultFeedbackRequest": {
        "properties": {
          "feedbackId": {
            "format": "int64",
            "type": "integer"
          },
          "value": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "required": [
          "feedbackId",
          "value"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MLSuite API",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/infrastructure/services/{serviceName}/actions": {
      "post": {
        "operationId": "action",
        "parameters": [
          {
            "in": "path",
            "name": "serviceName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "infrastructure-controller"
        ]
      }
    },
    "/api/admin/infrastructure/services/{serviceName}/logs": {
      "get": {
        "operationId": "logs",
        "parameters": [
          {
            "in": "path",
            "name": "serviceName",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tail",
            "required": false,
            "schema": {
              "default": 200,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonNode"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "infrastructure-controller"
        ]
      }
    },
    "/api/admin/infrastructure/terminal/sessions": {
      "post": {
        "operationId": "createTerminalSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalSessionResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "infrastructure-controller"
        ]
      }
    },
    "/api/admin/infrastructure/terminal/sessions/{sessionId}": {
      "delete": {
        "operationId": "closeTerminalSession",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "infrastructure-controller"
        ]
      }
    },
...
    "/api/plugins": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "plugin-controller"
        ]
      },
      "get": {
        "operationId": "getAll",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 24,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "all",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoPluginDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "plugin-controller"
        ]
      },
      "post": {
        "operationId": "upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "plugin-controller"
        ]
      }
    },
...
    "/api/prediction-result-feedback": {
      "get": {
        "operationId": "list_4",
        "parameters": [
          {
            "in": "query",
            "name": "resultId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PredictionResultFeedbackDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "prediction-result-feedback-controller"
        ]
      },
      "patch": {
        "operationId": "update_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePredictionResultFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResultFeedbackDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "prediction-result-feedback-controller"
        ]
      },
      "post": {
        "operationId": "create_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePredictionResultFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResultFeedbackDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "prediction-result-feedback-controller"
        ]
      }
    },
    "/api/prediction-result-feedback/by-runs": {
      "get": {
        "operationId": "listByRuns",
        "parameters": [
          {
            "in": "query",
            "name": "runIds",
            "required": true,
            "schema": {
              "items": {
                "format": "int64",
                "type": "integer"
              },
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PredictionResultFeedbackDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "prediction-result-feedback-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}